refactor(tags-bar): re-org dir structure for fold feat
