+++ src/containers/unit/TagsBar/styles/desktop_view/index.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components'
+
+import { css } from '@/utils'
+
+export const Wrapper = styled.div`
+  position: relative;
+  ${css.flexColumn()};
+  margin-top: 30px;
+  margin-left: 12px;
+`
+export const holder = 1
